chore: renamed openai fullstream examples
